[hir2mpl]Internal commit msg:     [DTS2022012712052] bugfix deal with 0xdeadbeef in lower and upper bound [mapleall]Internal commit msg:     [mapleall] pclint waring clean14     [mplcg] check cases for reconstructed patterns based-on ssa     [DTS2022012712052] bugfix deal with 0xdeadbeef in lower and upper bound

+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_dce.cpp
@@ -38,7 +38,7 @@ bool AArch64Dce::RemoveUnuseDef(VRegVersion &defVersion) {
       defVersion.MarkDeleted();
       uint32 opndNum = defInsn->GetOperandSize();
       for (int i = opndNum - 1; i >= 0; --i) {
-        Operand &opnd = defInsn->GetOperand(i);
+        Operand &opnd = defInsn->GetOperand(static_cast<uint32>(i));
         A64DeleteRegUseVisitor deleteUseRegVisitor(*GetSSAInfo(), defInsn->GetId());
         CHECK_FATAL(!opnd.IsPhi(), "unexpect phi insn");
         opnd.Accept(deleteUseRegVisitor);
